feat(dynamodb): allow optional indexes in constructor options
Updated the Dynamodb class to make the indexes parameter optional in the constructor. If no indexes are provided, the class will default to an empty array. This change improves the flexibility of the API, allowing users to instantiate the class without needing to specify indexes when they may not be necessary. Incremented version number to 1.0.2 in package.json to reflect this enhancement.